Corrigindo layout da activity_login.xml para melhorar suporte em telas pequenas: usando componentes menores e adicionando scroll global O email e senha que você digita no LoginFragment não é perdido ao se mudar para o SignInFragment Adicionando shelves no gitignore
