+++ .gitignore
@@ -2,6 +2,7 @@
 .gradle
 /local.properties
 /.idea/caches
+/.idea/shelf
 /.idea/libraries
 /.idea/modules.xml
 /.idea/workspace.xml
