--- .idea/shelf/ADN-44.xml
@@ -1,4 +1,0 @@
-<changelist name="ADN-44" date="1606343752729" recycled="false">
-  <option name="PATH" value="$PROJECT_DIR$/.idea/shelf/ADN-44/shelved.patch" />
-  <option name="DESCRIPTION" value="ADN-44" />
-</changelist>
