--- .idea/shelf/adn-38.xml
@@ -1,4 +1,0 @@
-<changelist name="adn-38" date="1606353881935" recycled="false" deleted="true">
-  <option name="PATH" value="$PROJECT_DIR$/.idea/shelf/adn-38/shelved.patch" />
-  <option name="DESCRIPTION" value="adn-38" />
-</changelist>
