+++ app/src/main/res/layout/user_email_password.xml
@@ -18,26 +18,22 @@
             android:layout_height="match_parent"
             android:layout_marginStart="10dp"
             android:contentDescription="@string/email"
-            app:srcCompat="@drawable/email_icon" />
+            app:srcCompat="@drawable/email_icon"/>
 
-        <com.google.android.material.textfield.TextInputLayout
+        <com.google.android.material.textfield.TextInputEditText
+            android:id="@+id/tiEmail"
             style="@style/Widget.MaterialComponents.TextInputLayout.OutlinedBox"
             app:boxStrokeColor="@null"
+            android:backgroundTint="@color/transparent"
             android:textColorHint="@color/black"
+            android:textColor="@color/black"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
             android:background="@color/transparent"
-            android:layout_marginEnd="5dp">
+            android:layout_marginEnd="5dp"
+            android:hint="@string/email"
+            android:inputType="textEmailAddress"/>
 
-            <com.google.android.material.textfield.TextInputEditText
-                android:id="@+id/tiEmail"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:hint="@string/email"
-                android:textColor="@color/black"
-                android:inputType="textEmailAddress"
-                android:background="@color/transparent"/>
-        </com.google.android.material.textfield.TextInputLayout>
     </LinearLayout>
 
     <LinearLayout
@@ -56,24 +52,20 @@
             android:contentDescription="@string/password"
             app:srcCompat="@drawable/password_icon" />
 
-        <com.google.android.material.textfield.TextInputLayout
+        <com.google.android.material.textfield.TextInputEditText
+            android:id="@+id/tiPassword"
             style="@style/Widget.MaterialComponents.TextInputLayout.OutlinedBox"
             app:boxStrokeColor="@null"
+            android:backgroundTint="@color/transparent"
             android:textColorHint="@color/black"
+            android:textColor="@color/black"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
             android:background="@color/transparent"
-            android:layout_marginEnd="5dp">
+            android:layout_marginEnd="5dp"
+            android:hint="@string/password"
+            android:inputType="textPassword"/>
 
-            <com.google.android.material.textfield.TextInputEditText
-                android:id="@+id/tiPassword"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:hint="@string/password"
-                android:textColor="@color/black"
-                android:inputType="textPassword"
-                android:background="@color/transparent"/>
-        </com.google.android.material.textfield.TextInputLayout>
     </LinearLayout>
 
 </LinearLayout>
